+++ codalab/worker/main.py
@@ -18,9 +18,9 @@ from codalab.lib.formatting import parse_size
 from .bundle_service_client import BundleServiceClient, BundleAuthException
 from . import docker_utils
 from .worker import Worker
-from .local_run.local_dependency_manager import LocalFileSystemDependencyManager
-from .local_run.docker_image_manager import DockerImageManager
-from .local_run.local_run_manager import LocalRunManager
+from dependency_manager import DependencyManager
+from docker_image_manager import DockerImageManager
+from run_manager import RunManager
 
 logger = logging.getLogger(__name__)
 
@@ -160,17 +160,17 @@ chmod 600 %s"""
         logging.debug('Work dir %s doesn\'t exist, creating.', args.work_dir)
         os.makedirs(args.work_dir, 0o770)
 
-    def create_local_run_manager(worker):
+    def create_run_manager(worker):
         """
         To avoid circular dependencies the Worker initializes takes a RunManager factory
-        to initilize its run manager. This method creates a LocalFilesystem-Docker RunManager
+        to initilize its run manager. This method creates a RunManager
         which is the default execution architecture Codalab uses
         """
         docker_runtime = docker_utils.get_available_runtime()
         cpuset = parse_cpuset_args(args.cpuset)
         gpuset = parse_gpuset_args(args.gpuset)
 
-        dependency_manager = LocalFileSystemDependencyManager(
+        dependency_manager = DependencyManager(
             os.path.join(args.work_dir, 'dependencies-state.json'),
             bundle_service,
             args.work_dir,
@@ -181,7 +181,7 @@ chmod 600 %s"""
             os.path.join(args.work_dir, 'images-state.json'), max_images_bytes
         )
 
-        return LocalRunManager(
+        return RunManager(
             worker,
             image_manager,
             dependency_manager,
@@ -195,7 +195,7 @@ chmod 600 %s"""
         )
 
     worker = Worker(
-        create_local_run_manager,
+        create_run_manager,
         os.path.join(args.work_dir, 'worker-state.json'),
         args.id,
         args.tag,
